Updated: Changed the name of getTop5UnsatissfiedStudents to getAllStudentsByUnsatisfactionRate
added functionality to getAllStudentsByUnsatisfactionRate - now the votes will be set in a particular order depending on the priority. (tested and works)

The method name is changed also in Interface, ControllerMock, GUI, Testing, so that everything can work with the new name.

Added functionality in facadeMock: 
1. Randomized the votes for every user, so that the new method could be fully tested.
2. Fixed the public UserType getUserTypeByName(String type) , now If statement "says" .equals instead of .endsWith
3. Update in public void addProposedSubject(ProposedSubject ps) method. Now every new Subject will receive a new id (this functionality was forgotten and I could have given a problem in the future)

all is checked and it's working.